+++ Editor/ObjectMeshDownloaderEditor.cs
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2025 MultiSet AI. All rights reserved.
+Copyright (c) 2026 MultiSet AI. All rights reserved.
 Licensed under the MultiSet License. You may not use this file except in compliance with the License. and you can’t re-distribute this file without a prior notice
 For license details, visit www.multiset.ai.
 Redistribution in source or binary forms must retain this notice.
@@ -10,14 +10,14 @@ using UnityEngine;
 
 namespace MultiSet
 {
-    [CustomEditor(typeof(ModelsetMeshDownloader))]
-    public class ModelSetMeshDownloaderEditor : Editor
+    [CustomEditor(typeof(ObjectMeshDownloader))]
+    public class ObjectMeshDownloaderEditor : Editor
     {
         public override void OnInspectorGUI()
         {
             DrawDefaultInspector();
 
-            ModelsetMeshDownloader mapMeshDownloader = (ModelsetMeshDownloader)target;
+            ObjectMeshDownloader mapMeshDownloader = (ObjectMeshDownloader)target;
 
             GUILayout.Space(20);
             EditorGUILayout.HelpBox("Download Object Mesh in Editor Mode. Helps to setup AR Scene", MessageType.Info);